Implement OAuth 2.0 refresh access token
Some OAuth providers ask to extend the access token expiration date, in
order to continue using it.
Microsoft Live do that, and it will be also soon mandatory for Facebook
because they will deprecate the 'offline_access' permission.

To refresh an access token, you need to access the accessTokenEndPoint
with other parameters like a specific 'grant_type' parameter.