distinguish between OAuth1 and OAuth2 tokens
The existing code was using the same base class for OAuth1 and OAuth2 request and access tokens. While nice for the API, this was a bit limiting for functionality. It was also confusing because OAuth2 does not have request tokens, and OAuth2 access tokens do not hold a secret.

I created the concept of RequestTokens and AccessTokens and refactored the code to be able to use the correct type depending on the API.

Backwards compatibility was preserved even though it makes the API a bit uglier.