Add MediaWiki API
This API can be used for any MediaWiki installation that has the OAuth
extension [1] installed. Two default instances are provided for
convenience: one for wikis hosted by the Wikimedia Foundation (including
Wikipedia), which is the one most users will be interested in, and one
for the Wikimedia Foundation’s Beta Cluster, which is more appropriate
for testing purposes.

The example is copied and slightly adapted from AWeberExample.java.

[1]: https://www.mediawiki.org/wiki/Extension:OAuth